+++ packages/shared/src/catalog/index.ts
@@ -10,6 +10,7 @@ export * from "./primitives/numberField";
 export * from "./primitives/searchField";
 export * from "./primitives/separator";
 export * from "./primitives/textField";
+export * from "./primitives/timeField";
 export * from "./primitives/toolbar";
 export * from "./primitives/toggleButton";
 export * from "./primitives/toggleButtonGroup";